fix a compiler warning.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@4083 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/ParallelPort/ParallelPort.cpp
@@ -275,7 +275,7 @@ int CParallelPort::OnControlRegister(MM::PropertyBase* pProp, MM::ActionType eAc
 	{
 		long pos;
 		pProp->Get(pos);
-      short buf = pos;
+      short buf = static_cast<short>(pos);
    	const short addr = 0x378 + 2; // the control register is the base port + 2
 	   Out32(addr, buf);
    }
